changing int to NSInteger

+++ TicTacToe/TicTacToe/Classes/model/TicTacToeBoard.h
@@ -9,7 +9,7 @@
 
 // Returns the state of the square at the given position on the board.
 // Invalid posions will return TicTacToeStateInvalid.
-- (TicTacToeStateType)stateForXPos:(int)x yPos:(int)y;
+- (TicTacToeStateType)stateForXPos:(NSInteger)x yPos:(NSInteger)y;
 
 // Attempts to play in the position (x,y). Returns YES for success, NO if not.
 - (BOOL)playXPos:(NSInteger)x
